adds page Sobre, moves navigation to component, and adds page title to latest posts
