+++ src/components/footer.tsx
@@ -1,59 +1,10 @@
-"use client";
-
-import Link from "next/link";
-import { usePathname } from "next/navigation";
+import Navigation from "./navigation";
 
 export default function Footer() {
-  const pathname = usePathname();
-  const activeClass =
-    "p-2 hover:underline hover:underline-offset-8 decoration-2 decoration-wavy decoration-blue-600 font-bold underline underline-offset-8";
-  const innactiveClass =
-    "p-2 hover:underline hover:underline-offset-8 decoration-2 decoration-wavy decoration-blue-600";
-
   return (
     <footer className="bg-white" id="footer-menu">
       <div className="mx-auto max-w-7xl overflow-hidden px-6 py-20 sm:py-24 lg:px-8">
-        <nav
-          className="-mb-6 columns-2 sm:flex sm:justify-center sm:space-x-12"
-          aria-label="Footer"
-        >
-          <div className="pb-6">
-            <Link
-              className={pathname === "/" ? activeClass : innactiveClass}
-              href="/"
-            >
-              Inicial
-            </Link>
-          </div>
-          <div className="pb-6">
-            <Link
-              className={
-                pathname.includes("/artigos") ? activeClass : innactiveClass
-              }
-              href="/artigos"
-            >
-              Artigos
-            </Link>
-          </div>
-          <div className="pb-6">
-            <Link
-              className={
-                pathname.includes("/projetos") ? activeClass : innactiveClass
-              }
-              href="/projetos"
-            >
-              Projetos
-            </Link>
-          </div>
-          <div className="pb-6">
-            <Link
-              className={pathname === "/contato" ? activeClass : innactiveClass}
-              href="/contato"
-            >
-              Contato
-            </Link>
-          </div>
-        </nav>
+        <Navigation className="max-auto" />
         <div className="mt-10 flex justify-center space-x-10">
           <a
             href="https://instagram.com/itsmeerickpatrick"
