+++ src/components/header.tsx
@@ -1,16 +1,6 @@
-"use client";
-
-import Link from "next/link";
-import { usePathname } from "next/navigation";
+import Navigation from "./navigation";
 
 export default function Header() {
-  const activeClass =
-    "p-2 hover:underline hover:underline-offset-8 decoration-2 decoration-wavy decoration-blue-600 font-bold underline underline-offset-8";
-  const innactiveClass =
-    "p-2 hover:underline hover:underline-offset-8 decoration-2 decoration-wavy decoration-blue-600";
-
-  const pathname = usePathname();
-
   return (
     <div className="p-4 flex justify-between align-center">
       <a href="/" className="group">
@@ -22,36 +12,8 @@ export default function Header() {
           </div>
         </div>
       </a>
-      <nav className="hidden sm:flex py-3 justify-between">
-        <Link
-          className={pathname === "/" ? activeClass : innactiveClass}
-          href="/"
-        >
-          Inicial
-        </Link>
-        <Link
-          className={
-            pathname.includes("/artigos") ? activeClass : innactiveClass
-          }
-          href="/artigos"
-        >
-          Artigos
-        </Link>
-        <Link
-          className={
-            pathname.includes("/projetos") ? activeClass : innactiveClass
-          }
-          href="/projetos"
-        >
-          Projetos
-        </Link>
-        <Link
-          className={pathname === "/contato" ? activeClass : innactiveClass}
-          href="/contato"
-        >
-          Contato
-        </Link>
-      </nav>
+
+      <Navigation className="hidden sm:flex" />
 
       <a
         className="sm:hidden flex p-4 underline underline-offset-8"
