+++ src/components/home/latests-posts.tsx
@@ -1,7 +1,10 @@
 export default function LaterstPosts() {
   return (
     <div>
-      <div className="bg-white">
+      <h3 className="text-center text-lg sm:text-xl md:text-2xl lg:text-3xl xl:text-4xl font-black">
+        Últimos artigos que escrevi
+      </h3>
+      <div className="bg-white py-8 sm:py-16">
         <div className="mx-auto grid max-w-5xl grid-cols-1 gap-x-8 gap-y-12 sm:gap-y-16 lg:grid-cols-2">
           <article className="mx-auto w-full max-w-2xl lg:mx-0 lg:max-w-lg">
             <time
@@ -35,7 +38,7 @@ export default function LaterstPosts() {
           </article>
 
           <div className="mx-auto w-full max-w-2xl border-t border-gray-900/10 pt-12 sm:pt-16 lg:mx-0 lg:max-w-none lg:border-t-0 lg:pt-0">
-            <div className="-my-12 divide-y divide-gray-900/10">
+            <div className="divide-y divide-gray-900/10">
               <article className="py-6 relative flex flex-wrap gap-x-3">
                 <time
                   dateTime="2024-02-26 11:54:37"
